revert legacy, update RCAIDE functionality
