+++ Legacy/trunk/S/Methods/Geometry/Two_Dimensional/Cross_Section/Airfoil/__init__.py
@@ -3,7 +3,7 @@
 # @ingroup Methods-Geometry-Two_Dimensional-Cross_Section
 
 from .compute_naca_4series        import compute_naca_4series 
-#from .compute_airfoil_properties  import compute_airfoil_properties
+from .compute_airfoil_properties  import compute_airfoil_properties
 from .import_airfoil_dat          import import_airfoil_dat
 from .import_airfoil_geometry     import import_airfoil_geometry 
 from .import_airfoil_polars       import import_airfoil_polars
