+++ Legacy/trunk/S/Methods/Propulsion/__init__.py
@@ -4,6 +4,7 @@
 
 from . import Rotor_Wake
 from .ducted_fan_sizing import ducted_fan_sizing
+from .propeller_design import propeller_design
 from .turbofan_emission_index import turbofan_emission_index
 from .electric_motor_sizing import size_from_kv, size_from_mass
 from .turbofan_sizing import turbofan_sizing
