+++ Legacy/trunk/S/Plots/Geometry/__init__.py
@@ -3,4 +3,6 @@
 # @ingroup Plots
 
 from .plot_airfoil                  import plot_airfoil
+from .plot_vehicle                  import plot_vehicle
+from .plot_propeller                import plot_propeller
 from .plot_vehicle_vlm_panelization import plot_vehicle_vlm_panelization
