+++ RCAIDE/Methods/Geometry/Two_Dimensional/Cross_Section/Airfoil/__init__.py
@@ -10,8 +10,6 @@
 # ----------------------------------------------------------------------------------------------------------------------
 # @ingroup Methods-Geometry-Two_Dimensional-Cross_Section
 
-from Legacy.trunk.S.Methods.Geometry.Two_Dimensional.Cross_Section.Airfoil import compute_naca_4series 
-#from Legacy.trunk.S.Methods.Geometry.Two_Dimensional.Cross_Section.Airfoil import compute_airfoil_properties
 from Legacy.trunk.S.Methods.Geometry.Two_Dimensional.Cross_Section.Airfoil import import_airfoil_dat
 from Legacy.trunk.S.Methods.Geometry.Two_Dimensional.Cross_Section.Airfoil import import_airfoil_geometry 
 from Legacy.trunk.S.Methods.Geometry.Two_Dimensional.Cross_Section.Airfoil import import_airfoil_polars
